# chore(Travis): Speed up travis builds and deploys by reducing babel p…