--- a/.gitignore
+++ b/.gitignore
@@ -1,6 +1,7 @@
 dist
 lib
 static
+tmp
 node_modules
 npm-debug.log
 yarn-error.log