…asses

Jest on Travis is super slow and sometimes times out. This is a test to see if running babel first separately speeds things up. It also uses the babel stuff for prepublishing. 

Also I've added `build:static` to the script section to highlight any issues with the state site generation in PRs. 

Local test runs. 

```
# Running Jest with Babel
[1] - 45.746s
[2] - 41.68s
[3] - 40.597s

# Running Jest on pre-babelified (times include babel run)
[1] - 25.3s
[2] - 27.88s
[3] - 27.516s
```