@HHogg 2017-05-12
Ok this seems to have drastically improved things

```
# Travis running Jest with Babel
[1] - 727.282s
[2] - 495.729s
[3] - 588.327s

# Travis now running Jest on pre-babelified
[1] - 7.167s
[2] - 9.719s
[3] - 5.867s
```

Overall it's cut the build time from around **11m 33s*** to **3m 9s*** (**~360% improvement!**)

_* Taking the last 3 builds as an average_